% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parsers.R
\name{parse_FORTYPCD}
\alias{parse_FORTYPCD}
\title{Parse FORTYPCD group strings into code and name}
\usage{
parse_FORTYPCD(x)
}
\arguments{
\item{x}{Character vector of raw group strings, e.g.
\code{"`0161 Loblolly pine"}.}
}
\value{
A data frame with columns \code{FORTYPCD}, \code{FORTYPCD_NAME}.
}
\description{
Parse FORTYPCD group strings into code and name
}
\keyword{internal}
